Split cdcc.py into one file per variable
