+++ openfisca_us/variables/irs/credits/cdcc/cdcc_refund.py
@@ -1,0 +1,17 @@
+from openfisca_us.model_api import *
+
+
+class cdcc_refund(Variable):
+    value_type = float
+    entity = TaxUnit
+    definition_period = YEAR
+    label = "Child/dependent care refundable credit"
+    unit = USD
+    documentation = "Refundable credit for child and dependent care expenses from Form 2441"
+
+    def formula(tax_unit, period, parameters):
+        cdcc = parameters(period).irs.credits.cdcc
+        if cdcc.refundable and not cdcc.abolition:
+            return tax_unit("c33200", period)
+        else:
+            return 0
